+++ src/MatrixMultiplicationMap2.java
@@ -1,11 +1,12 @@
 import java.io.IOException;
-import org.apache.hadoop.io.IntWritable;
+
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
 
-public class MatrixMultiplicationMap2 extends Mapper<LongWritable, Text, Text, IntWritable> 
+public class MatrixMultiplicationMap2 extends Mapper<LongWritable, Text, Text, FloatWritable> 
 {	
 	public void map(LongWritable offset,Text line,Context context) throws IOException,InterruptedException
 	{
@@ -14,9 +15,9 @@ public class MatrixMultiplicationMap2 extends Mapper<LongWritable, Text, Text, I
 		int starthash = record.indexOf("###");
 		int row = Integer.parseInt(record.substring(0, startdollar).trim());
 		int col = Integer.parseInt(record.substring(startdollar+3, starthash).trim());
-		int product = Integer.parseInt(record.substring(starthash+3).trim());
+		float product = Float.parseFloat(record.substring(starthash+3).trim());
 		
 		System.out.println("Now writing row: "+row+" $$$ "+col+" "+product);
-		context.write(new Text(row+" $$$ "+col), new IntWritable(product));
+		context.write(new Text(row+" $$$ "+col), new FloatWritable(product));
 	}
 }
